# services/common/src/main/java/com/gigshield/common/dto/AuthRequest.java
package com.gigshield.common.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

/**
 * DTO for authentication request.
 */
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class AuthRequest {
    
    @NotBlank(message = "Email or phone is required")
    private String identifier;  // Email or phone
    
    @NotBlank(message = "Password is required")
    private String password;
}